modules/robin/requirements.txt
summ-web<=0.1.0
summ-logging<=0.1.0
hawk-models<=0.1.0
robin-core<=0.1.0
Flask # version managed by summ-web
Flask-RESTful # version managed by summ-web
Flask-SQLAlchemy # version managed by summ-web
Flask-Cors # version managed by summ-web
webargs # version managed by summ-web
